requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
tiktoken
requests
pdfplumber
python-docx
sentence-transformers
